pyproject.toml
[tool.poetry]
name = "SegAndDiffuse"
version = "0.1.0"
description = ""
authors = ["David Wessels <davidwessels15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
numpy = "^1.24.1"
pytest = "^7.2.0"
coverage = "^7.0.3"
coverage-badge = "^1.1.0"
pandas = "^1.5.3"
matplotlib = "^3.6.3"
ipython = "^8.14.0"
torch = "2.0.1"
torchvision = "0.15.2"
addict = "^2.4.0"
diffusers = "^0.17.1"
gradio = "^3.35.2"
huggingface-hub = "^0.15.1"
onnxruntime = "^1.15.1"
opencv-python = "^4.7.0.72"
pillow = "^9.5.0"
pycocotools = "^2.0.6"
pyyaml = "^6.0"
requests = "^2.31.0"
setuptools = "^67.8.0"
supervision = "^0.10.0"
termcolor = "^2.3.0"
timm = "^0.9.2"
transformers = "^4.30.2"
yapf = "^0.40.0"
nltk = "^3.8.1"
fairscale = "^0.4.13"
accelerate = "^0.20.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
